requirements.txt
numpy
tensorflow
Pillow
tqdm
seaborn

